update mutations, remove remaining refs
