+++ apps/web/src/pages/Applications/ApplicationCareerTimelineAddPage/components/AddExperienceForm.tsx
@@ -8,10 +8,13 @@ import { Button, Heading, Link, Separator } from "@gc-digital-talent/ui";
 import { Select } from "@gc-digital-talent/forms";
 import { errorMessages, formMessages } from "@gc-digital-talent/i18n";
 import { useAuthorization } from "@gc-digital-talent/auth";
+import { Scalars } from "@gc-digital-talent/graphql";
 
 import useRoutes from "~/hooks/useRoutes";
-import { useExperienceMutations } from "~/hooks/useExperienceMutations";
-import { Scalars } from "~/api/generated";
+import {
+  useExperienceMutations,
+  isSuccessfulCreate,
+} from "~/hooks/useExperienceMutations";
 import {
   formValuesToSubmitData,
   getExperienceFormLabels,
@@ -26,7 +29,6 @@ import ExperienceDetails from "~/components/ExperienceFormFields/ExperienceDetai
 import ErrorSummary from "~/components/ExperienceFormFields/ErrorSummary";
 
 import { experienceTypeTitles } from "../messages";
-import { isSuccessfulCreate } from "./addExperienceUtils";
 
 type FormAction = "return" | "add-another";
 type ExperienceExperienceFormValues =
@@ -35,7 +37,7 @@ type ExperienceExperienceFormValues =
     action: FormAction | "";
   };
 interface AddExperienceFormProps {
-  applicationId: Scalars["ID"];
+  applicationId: Scalars["ID"]["output"];
 }
 
 const AddExperienceForm = ({ applicationId }: AddExperienceFormProps) => {
