+++ apps/web/src/pages/Applications/ApplicationEducationPage/utils.tsx
@@ -5,8 +5,8 @@ import uniqueId from "lodash/uniqueId";
 import { Locales } from "@gc-digital-talent/i18n";
 import { Link } from "@gc-digital-talent/ui";
 import { Radio } from "@gc-digital-talent/forms";
+import { EducationRequirementOption } from "@gc-digital-talent/graphql";
 
-import { EducationRequirementOption } from "~/api/generated";
 import applicationMessages from "~/messages/applicationMessages";
 import { ClassificationGroup } from "~/utils/poolUtils";
