+++ apps/web/src/pages/Applications/ApplicationSelfDeclarationPage/SelfDeclaration/SelfDeclarationForm.stories.tsx
@@ -4,8 +4,7 @@ import { action } from "@storybook/addon-actions";
 
 import { CHROMATIC_VIEWPORTS } from "@gc-digital-talent/storybook-helpers";
 import { fakePoolCandidates } from "@gc-digital-talent/fake-data";
-
-import { IndigenousCommunity } from "~/api/generated";
+import { IndigenousCommunity } from "@gc-digital-talent/graphql";
 
 import { ApplicationSelfDeclaration } from "../ApplicationSelfDeclarationPage";
