+++ apps/web/src/pages/Applications/applicationOperations.graphql
@@ -118,29 +118,6 @@ mutation UpdateApplication($id: ID!, $application: UpdateApplicationInput!) {
   }
 }
 
-mutation SubmitApplication($id: ID!, $signature: String!) {
-  submitApplication(id: $id, signature: $signature) {
-    id
-    signature
-  }
-}
-
-mutation UpdateUserAndApplication(
-  $userId: ID!
-  $userInput: UpdateUserAsUserInput!
-  $applicationId: ID!
-  $applicationInput: UpdateApplicationInput!
-) {
-  updateUserAsUser(id: $userId, user: $userInput) {
-    id
-    indigenousCommunities
-    indigenousDeclarationSignature
-  }
-  updateApplication(id: $applicationId, application: $applicationInput) {
-    id
-  }
-}
-
 fragment experienceSubTypes on Experience {
   ... on AwardExperience {
     title
