+++ apps/web/src/pages/Applications/questionsStep/questionsStepValidation.tsx
@@ -1,4 +1,5 @@
-import { User, Pool, PoolCandidate } from "~/api/generated";
+import { User, Pool, PoolCandidate } from "@gc-digital-talent/graphql";
+
 import { screeningQuestionsSectionHasMissingResponses } from "~/validators/profile";
 
 const stepHasError = (
