+++ apps/web/src/pages/Applications/reviewStep/reviewStepValidation.tsx
@@ -1,4 +1,4 @@
-import { User, Pool, PoolCandidate } from "~/api/generated";
+import { User, Pool, PoolCandidate } from "@gc-digital-talent/graphql";
 
 const stepHasError = (
   _user: User,
