+++ apps/web/src/pages/Applications/skillsStep/skillsStepValidation.tsx
@@ -1,4 +1,5 @@
-import { User, Pool } from "~/api/generated";
+import { User, Pool } from "@gc-digital-talent/graphql";
+
 import { skillRequirementsIsIncomplete } from "~/validators/profile";
 
 const stepHasError = (user: User, pool: Pool) => {
